refactor : 성능 개선
