+++ build.gradle
@@ -39,8 +39,8 @@ spotless {
 	}
 }
 
-dependencies {
 
+dependencies {
 
 	// jpa
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
@@ -75,10 +75,14 @@ dependencies {
 	// prometheus
 	runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
 
+	// caffein cache
+	implementation 'org.springframework.boot:spring-boot-starter-cache'
+	implementation("com.github.ben-manes.caffeine:caffeine:3.1.8")
 
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.security:spring-security-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+
 }
 
 tasks.named('test') {
