+++ src/test/java/com/example/solo/schedule/application/facade/AddScheduleFacadeTest.java
@@ -1,11 +1,6 @@
-package com.example.solo.schedule.facade;
+package com.example.solo.schedule.application.facade;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.List;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -21,17 +16,15 @@ import com.example.solo.member.domain.encrypt.PasswordEncryptor;
 import com.example.solo.member.domain.entity.Member;
 import com.example.solo.member.domain.entity.Password;
 import com.example.solo.member.domain.repository.MemberRepository;
-import com.example.solo.schedule.application.facade.GetMonthScheduleFacade;
-import com.example.solo.schedule.domain.dto.response.GetScheduleResponse;
-import com.example.solo.schedule.domain.entity.Schedule;
+import com.example.solo.schedule.domain.dto.request.AddScheduleRequestDto;
 import com.example.solo.schedule.domain.repository.ScheduleRepository;
 
 @ActiveProfiles("test")
 @SpringBootTest
 @AutoConfigureMockMvc
-public class GetMonthScheduleFacadeTest {
+public class AddScheduleFacadeTest {
 
-  @Autowired private GetMonthScheduleFacade getMonthScheduleFacade;
+  @Autowired private AddScheduleFacade addScheduleFacade;
   @Autowired private MemberRepository memberRepository;
   @Autowired private ScheduleRepository scheduleRepository;
 
@@ -58,26 +51,16 @@ public class GetMonthScheduleFacadeTest {
             .build();
 
     memberRepository.save(member);
-
-    scheduleRepository.save(
-        Schedule.builder()
-            .member(member)
-            .category(TEST_CATEGORY)
-            .date(LocalDate.now())
-            .time(LocalTime.of(TEST_HOUR, TEST_MINUTES))
-            .build());
   }
 
   @Test
-  @DisplayName("일정을 조회한다.")
-  void getMonthScheduleSuccess() {
-    // when
-    List<GetScheduleResponse> result =
-        getMonthScheduleFacade.getSchedules(
-            member, LocalDate.now().getYear(), LocalDate.now().getMonthValue());
+  @DisplayName("일정을 생성한다.")
+  void addSchedule() {
+    // given
+    AddScheduleRequestDto requestDto =
+        new AddScheduleRequestDto(TEST_CATEGORY, TEST_HOUR, TEST_MINUTES);
 
-    // then
-    assertEquals(1, result.size());
-    assertTrue(result.stream().anyMatch(s -> s.category().equals(TEST_CATEGORY)));
+    // when + then
+    assertDoesNotThrow(() -> addScheduleFacade.addSchedule(member, requestDto));
   }
 }
